backend/requirements.txt
starlette==0.41.1
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.26.0
redis==5.0.1
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
pytest==7.4.4
